third_party/blink/web_tests/fast/css/forced-colors-mode/forced-colors-mode-04.html

<!DOCTYPE html>
<html>
<head>
<title>Forced colors mode - webkit-tap-highlight-color.</title>
<style type="text/css">
  a {
    -webkit-tap-highlight-color: rgb(0, 0, 255);
  }
</style>
</head>
<body>
  <p>
    <a href="https://www.wikipedia.org" id="link">
      This link color should be overridden when forced colors mode is enabled.
    </a>
  </p>
</body>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
  test(function(){
    assert_equals(getComputedStyle(link).webkitTapHighlightColor, "rgba(0, 0, 0, 0.18)");
  }, "Checks hyperlinks are overridden in forced colors mode.");
</script>
</html>
